Mark zeroconf.handlers as protected by renaming to zeroconf._handlers
- The public API should only access zeroconf and zeroconf.aio
  as internals may be relocated between releases
