+++ zeroconf/_core.py
@@ -29,6 +29,7 @@ from types import TracebackType  # noqa # used in type hints
 from typing import Dict, List, Optional, Type, Union, cast
 
 from ._dns import DNSIncoming, DNSOutgoing, DNSQuestion
+from ._handlers import QueryHandler, RecordManager
 from .cache import DNSCache
 from .const import (
     _CACHE_CLEANUP_INTERVAL,
@@ -47,7 +48,6 @@ from .const import (
     _UNREGISTER_TIME,
 )
 from .exceptions import NonUniqueNameException
-from .handlers import QueryHandler, RecordManager
 from .logger import QuietLogger, log
 from .services import (
     RecordUpdateListener,
